Added game title, wrapper, placeholder window, player panel, and a default font
