+++ src/App.css
@@ -1,3 +1,7 @@
+html {
+    font-family: Verdana, Geneva, Tahoma, sans-serif;
+}
+
 .wrapper {
     padding: 20px;
 }
